+++ companion-module-taisto/webpack.config.cjs
@@ -1,0 +1,6 @@
+const createCompanionConfig = require("@companion-module/tools/webpack.config.cjs");
+
+module.exports = (env = {}, argv = {}) => createCompanionConfig(
+  Object.assign({}, env, { ROOT: __dirname }),
+  argv
+);