+++ js/routes.js
@@ -18,6 +18,7 @@ import TranslationList from "./containers/TranslationList";
 import Help from "./components/Help";
 import ConGroupList from "./components/ConGroupList";
 import ApiKeySettings from "./components/ApiKeySettings";
+import AuditLogs from "./components/AuditLogs";
 
 export default function Routes() {
   return (
@@ -35,6 +36,7 @@ export default function Routes() {
       <Route exact path="/settings/oletustilat" component={DefaultStatesSettingsList} />
 	  <Route exact path="/settings/output-groups" component={ConGroupList} />
 	  <Route exact path="/settings/api-key" component={ApiKeySettings} />
+	  <Route exact path="/settings/audit-logs" component={AuditLogs} />
       <Route exact path="/settings/diagram/:slug" component={DiagramSettings} />
       <Route exact path="/settings/diagrams" component={DiagramList} />
       <Route exact path="/settings/translations" component={TranslationList} />